refactor: split LspServer into focused components
LspServer.cs owned transport framing, JSON-RPC binding, workspace and
document overlay state, validation scheduling, diagnostic publishing,
UTF-16 mapping, navigation/completion/hover/rename/code actions/code
lenses, URI conversion, and error handling in one 891-line class.

Split it into single-responsibility components, behavior unchanged:

- Logger.cs: leveled stderr logging
- LspTransport.cs: Content-Length framing, notifications, FrameHandler
- WorkspaceState.cs: root, document overlay, engine/symbol lifecycle,
  URI/path conversion, Utf16Mapper cache
- LspConversions.cs: engine <-> LSP position/range/diagnostic mapping
- DiagnosticsPublisher.cs: debounced validation and deduped publishing
- LanguageFeatures.cs: hover, navigation, completion, rename, code
  actions, code lenses
- LspServer.cs: JSON-RPC target, lifecycle, and wiring only
